--- src/main/java/com/ddudu/infrastructure/persistence/adapter/SavePeriodGoalPersistenceAdapter.java
@@ -1,22 +1,0 @@
-package com.ddudu.infrastructure.persistence.adapter;
-
-import com.ddudu.application.domain.period_goal.domain.PeriodGoal;
-import com.ddudu.application.port.out.period_goal.SavePeriodGoalPort;
-import com.ddudu.infrastructure.annotation.DrivenAdapter;
-import com.ddudu.infrastructure.persistence.entity.PeriodGoalEntity;
-import com.ddudu.infrastructure.persistence.repository.period_goal.PeriodGoalRepository;
-import lombok.RequiredArgsConstructor;
-
-@DrivenAdapter
-@RequiredArgsConstructor
-public class SavePeriodGoalPersistenceAdapter implements SavePeriodGoalPort {
-
-  private final PeriodGoalRepository periodGoalRepository;
-
-  @Override
-  public PeriodGoal save(PeriodGoal periodGoal) {
-    return periodGoalRepository.save(PeriodGoalEntity.from(periodGoal))
-        .toDomain();
-  }
-
-}
